requirements.txt
customtkinter
Pillow
requests
packaging